added finetune script
add the imagenet data folder to the highest level of the repo. Then, add subfolders for bboxes_annotations and images. Insert the annotations in the first folder, and add the train and val folders in the images folder. pip install all the packages in ronak_learning/requirements.txt, and run the finetune script.